Implementing tree-sitter based indentation logic
- previously developed and tested in the sane-indentation package (> 0.9).

Refer to https://github.com/atom/language-javascript/issues/594#issuecomment-420415412

By itself this does nothing. The new logic is only used if the language package for the current language contains the necessary configuration (e.g., which scopes to indent on). So this PR goes together with, e.g., FILL-ME-IN in language-javascript.

Updated: now without the need for 'precedingRowCondition' callbacks in the languages-specific configuration